+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Changes
 * [UI]: Added html viewer for analysis output files.
 * [REST]: Added new REST endpoint that can generate a deep hashcode of a project.  This will be used for speeding up project synchronization when there are no changes to a project.  Instead of checking for changes remotely, the client can check for changes only to the hash.  If there are hash changes, then they can do their full sync.
 * [UI]: Added option for project managers to check for changes using the hash above, or to force a full project sync on the project remote settings page.
+* [Developer/REST/UI]: Updated announcements with a new title and priority fields.
 * [UI]: Added option for user to only receive an email upon pipeline failure.
 
 20.05 to 20.09